switch from pipe to named temp file. move parse_entity to utilities
